fixed some type collisions

+++ metrics_tf.py
@@ -55,8 +55,8 @@ def voxel_false_positive_error(voxel_gt, voxel_pred):
     # formula is false positive / (false positive + true negative)
     # TN = true negative means voxel_pred == 0 & voxel_gt == 0
     # FP = false positive means voxel_pred == 1 & voxel_gt == 0
-    tn = tf.reduce_sum(is_free(voxel_gt) & is_free(voxel_pred))
-    fp = tf.reduce_sum(is_obstacle(voxel_gt) & is_free(voxel_pred))
+    tn = tf.reduce_sum(tf.cast(is_free(voxel_gt) & is_free(voxel_pred), dtype=tf.int32))
+    fp = tf.reduce_sum(tf.cast(is_obstacle(voxel_gt) & is_free(voxel_pred), dtype=tf.int32))
     return fp / (fp + tn)
 
 
@@ -65,8 +65,8 @@ def voxel_true_positive_error(voxel_gt, voxel_pred):
     # formula is true positive / (false negative + true positive)
     # FN = false negative means voxel_pred == 0 & voxel_gt == 1
     # TP = true positive means voxel_pred == 1 & voxel_gt == 1
-    fn = tf.reduce_sum(is_free(voxel_gt) & is_obstacle(voxel_pred))
-    tp = tf.reduce_sum(is_obstacle(voxel_gt) & is_obstacle(voxel_pred))
+    fn = tf.reduce_sum(tf.cast(is_free(voxel_gt) & is_obstacle(voxel_pred), dtype=tf.int32))
+    tp = tf.reduce_sum(tf.cast(is_obstacle(voxel_gt) & is_obstacle(voxel_pred), dtype=tf.int32))
     return tp / (fn + tp)
 
 
@@ -74,7 +74,7 @@ def voxel_iou_error(voxel_gt, voxel_pred):
     # https://arxiv.org/pdf/1604.00449.pdf
     obst_pred = is_obstacle(voxel_pred)
     obst_gt = is_obstacle(voxel_gt)
-    return tf.reduce_sum(obst_gt & obst_pred) / tf.reduce_sum(obst_gt | obst_pred)
+    return tf.reduce_sum(tf.cast(obst_gt & obst_pred, dtype=tf.float32)) / tf.reduce_sum(tf.cast(obst_gt | obst_pred, dtype=tf.float32))
 
 
 def voxel_l1_dist_with_unknown(voxel_gt, voxel_pred):
